Clean up project
Add errors to interface methods to give callers more insight as to the
result of the operation.

Added a method to send a message to one consumer, and updated the existing SendMessage method to be a publish message which sends the message to all matching consumers.

Updated tests to match the changes to production code.

Fixed a few typos

Signed-off-by: Anthony M. Bonafide <AnthonyMBonafide@gmail.com>